requirements.txt
Django==2.0.3
django-autoslug-iplweb==1.9.4.dev0
django-cors-headers==2.2.0
django-redis==4.9.0
django-rest-framework==0.1.0
django-seed==0.1.8
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
flake8==3.5.0
ipython==6.2.1
psycopg2==2.7.4
pyflakes==1.6.0
python-dateutil==2.7.0
pytz==2018.3
